A few bugs have been fixed

+++ src/cron_trigger.c
@@ -230,7 +230,7 @@ unsigned int flag_starting_jobs(cron_line_t *cl)
 }
 
 //--------------------------------------------
-cron_trigger_t *cron_trigger_add_new(cron_trigger_t **list, const char *str, size_t next_id)
+cron_trigger_t *cron_trigger_add_new(cron_trigger_t **list, const char *str, uint32_t next_id)
 {
 	cron_trigger_t *item;
